# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIhidalgo.devs",
        "description": "Api escuela hidalgo",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "Equipo de soporte",
            "email": "msolegario.cisneros@gmail.com"
        },
        "license": {
            "name": "Licencia",
            "url": "https://example.com/license"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://por_definir/api/documentation"
        }
    ],
    "paths": {
...
        "/api/schedules": {
            "post": {
                "tags": [
                    "schedules"
                ],
                "summary": "Create a new schedule for an activity",
                "description": "Create a new schedule for a given activity, validating date and time, teacher availability, and activity schedule conflicts.",
                "operationId": "createScheduleForActivity",
                "requestBody": {
                    "description": "Schedule data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "activity_id",
                                    "teacher_id",
                                    "start_hour",
                                    "end_hour"
                                ],
                                "properties": {
                                    "activity_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "teacher_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_hour": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-06-10 08:00:00"
                                    },
                                    "end_hour": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-06-10 09:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Schedule created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "activity_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "teacher_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "start_hour": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-06-10 08:00:00"
                                        },
                                        "end_hour": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-06-10 09:00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "bad request"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "date invalid"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error creating schedule"
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Internal server error message"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/schedule/teacher/{id}": {
            "get": {
                "tags": [
                    "schedules"
                ],
                "summary": "Get schedule for a specific teacher by ID",
                "description": "Retrieve the schedule for a teacher specified by their ID.",
                "operationId": "getScheduleForTeacherId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the teacher",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "course": {
                                                        "type": "string",
                                                        "example": "programacion 1"
                                                    },
                                                    "activity": {
                                                        "type": "string",
                                                        "example": "ver video"
                                                    },
                                                    "start_hour": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-06-10 08:00:00"
                                                    },
                                                    "end_hour": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-06-10 09:00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Teacher not found or has no schedule",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "teacher has not schedule"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}